+++ 7382/GolovinaES/course_work/script.sh
